chore: add support for system label application rules (#235)

+++ label-application-rule-config-service-impl/src/main/java/org/hypertrace/label/application/rule/config/service/LabelApplicationRuleConfigServiceImpl.java
@@ -1,10 +1,12 @@
 package org.hypertrace.label.application.rule.config.service;
 
-import com.typesafe.config.Config;
 import io.grpc.Channel;
 import io.grpc.Status;
+import io.grpc.StatusRuntimeException;
 import io.grpc.stub.StreamObserver;
 import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import org.hypertrace.config.objectstore.ConfigObject;
@@ -27,29 +29,17 @@ import org.hypertrace.label.application.rule.config.service.v1.UpdateLabelApplic
 
 public class LabelApplicationRuleConfigServiceImpl
     extends LabelApplicationRuleConfigServiceGrpc.LabelApplicationRuleConfigServiceImplBase {
-  static final String LABEL_APPLICATION_RULE_CONFIG_SERVICE_CONFIG =
-      "label.application.rule.config.service";
-  static final String MAX_DYNAMIC_LABEL_APPLICATION_RULES_PER_TENANT =
-      "max.dynamic.label.application.rules.per.tenant";
-  static final int DEFAULT_MAX_DYNAMIC_LABEL_APPLICATION_RULES_PER_TENANT = 100;
   private final IdentifiedObjectStore<LabelApplicationRule> labelApplicationRuleStore;
   private final LabelApplicationRuleValidator requestValidator;
-  private final int maxDynamicLabelApplicationRulesAllowed;
+  private final LabelApplicationRuleConfig labelApplicationRuleConfig;
 
   public LabelApplicationRuleConfigServiceImpl(
-      Channel configChannel, Config config, ConfigChangeEventGenerator configChangeEventGenerator) {
-    int maxDynamicRules = DEFAULT_MAX_DYNAMIC_LABEL_APPLICATION_RULES_PER_TENANT;
-    if (config.hasPath(LABEL_APPLICATION_RULE_CONFIG_SERVICE_CONFIG)) {
-      Config labelApplicationRuleConfig =
-          config.getConfig(LABEL_APPLICATION_RULE_CONFIG_SERVICE_CONFIG);
-      if (labelApplicationRuleConfig.hasPath(MAX_DYNAMIC_LABEL_APPLICATION_RULES_PER_TENANT)) {
-        maxDynamicRules =
-            labelApplicationRuleConfig.getInt(MAX_DYNAMIC_LABEL_APPLICATION_RULES_PER_TENANT);
-      }
-    }
-    this.maxDynamicLabelApplicationRulesAllowed = maxDynamicRules;
+      Channel configChannel,
+      LabelApplicationRuleConfig labelApplicationRuleConfig,
+      ConfigChangeEventGenerator configChangeEventGenerator) {
+    this.labelApplicationRuleConfig = labelApplicationRuleConfig;
 
-    ConfigServiceBlockingStub configServiceBlockingStub =
+    final ConfigServiceBlockingStub configServiceBlockingStub =
         ConfigServiceGrpc.newBlockingStub(configChannel)
             .withCallCredentials(
                 RequestContextClientCallCredsProviderFactory.getClientCallCredsProvider().get());
@@ -96,9 +86,18 @@ public class LabelApplicationRuleConfigServiceImpl
           this.labelApplicationRuleStore.getAllObjects(requestContext).stream()
               .map(ConfigObject::getData)
               .collect(Collectors.toUnmodifiableList());
+      Set<String> labelApplicationRuleIds =
+          labelApplicationRules.stream()
+              .map(LabelApplicationRule::getId)
+              .collect(Collectors.toUnmodifiableSet());
+      List<LabelApplicationRule> filteredSystemLabelApplicationRules =
+          this.labelApplicationRuleConfig.getSystemLabelApplicationRules().stream()
+              .filter(rule -> !labelApplicationRuleIds.contains(rule.getId()))
+              .collect(Collectors.toUnmodifiableList());
       responseObserver.onNext(
           GetLabelApplicationRulesResponse.newBuilder()
               .addAllLabelApplicationRules(labelApplicationRules)
+              .addAllLabelApplicationRules(filteredSystemLabelApplicationRules)
               .build());
       responseObserver.onCompleted();
     } catch (Exception e) {
@@ -116,6 +115,12 @@ public class LabelApplicationRuleConfigServiceImpl
       LabelApplicationRule existingRule =
           this.labelApplicationRuleStore
               .getData(requestContext, request.getId())
+              .or(
+                  () ->
+                      Optional.ofNullable(
+                          this.labelApplicationRuleConfig
+                              .getSystemLabelApplicationRulesMap()
+                              .get(request.getId())))
               .orElseThrow(Status.NOT_FOUND::asRuntimeException);
       LabelApplicationRule updateLabelApplicationRule =
           existingRule.toBuilder().setData(request.getData()).build();
@@ -140,6 +145,14 @@ public class LabelApplicationRuleConfigServiceImpl
     try {
       RequestContext requestContext = RequestContext.CURRENT.get();
       this.requestValidator.validateOrThrow(requestContext, request);
+      String labelApplicationRuleId = request.getId();
+      if (this.labelApplicationRuleConfig
+          .getSystemLabelApplicationRulesMap()
+          .containsKey(labelApplicationRuleId)) {
+        // Deleting a system label application rule is not allowed
+        responseObserver.onError(new StatusRuntimeException(Status.INVALID_ARGUMENT));
+        return;
+      }
       this.labelApplicationRuleStore
           .deleteObject(requestContext, request.getId())
           .orElseThrow(Status.NOT_FOUND::asRuntimeException);
@@ -161,7 +174,8 @@ public class LabelApplicationRuleConfigServiceImpl
                   .filter(
                       action -> action.hasDynamicLabelExpression() || action.hasDynamicLabelKey())
                   .count();
-      if (dynamicLabelApplicationRules >= maxDynamicLabelApplicationRulesAllowed) {
+      if (dynamicLabelApplicationRules
+          >= this.labelApplicationRuleConfig.getMaxDynamicLabelApplicationRulesAllowed()) {
         throw Status.RESOURCE_EXHAUSTED.asRuntimeException();
       }
     }
